Move clear dates button to separate component
